BAU: Rename function to be more generic and therefore reusable
This will be used it multiple places outside of taxCredits
